+++ llvm/lib/ExecutionEngine/JITLink/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_llvm_component_library(LLVMJITLink
+  DWARFRecordSectionSplitter.cpp
   EHFrameSupport.cpp
   JITLink.cpp
   JITLinkGeneric.cpp
